V4-7/5/25
Continued to work on developing the front-end aspect by working on creating the side navigation bar. I decided try implement a navigation bar that when the screen size is smaller it turns into a top nav bar instead, this can improve the user experience a side bar can be hard to use on smaller screen sizes.