V5-15/5/25
I completed creating a functional side toggle bar with HTML, CSS and JavaScript. Although navigation bar was supposed to be on the side and when the screen size changes it moved to the top it was changed as it was not an essential feature. There were some challenges encountered which were that sometimes the naming across the different files wouldn't be correct and the navigation bar wouldn't work so I would need to go over them again to identify and inconsistencies with the naming. There was a feature added so that when the navigation bar is collapsed the arrow flips sides to represent expanding the navigation bar, although it is a very simple feature it can help improve the user interface as it makes the design more logical.