V6-21/5/25
I created a HTML and CSS page for the Profile section and also created a HTML page for on of the subject pages which is English. More changes were made to the homepage and a dropdown menu for the subjects section was attempted to be created however, it hasn't been completed yet.