V7- 1/6/25
Continued to work on creating the cards and a drop down menu for side navigation bar in the home page however, it could not be completed as there some errors where the card and side navigation bar kept overlapping which still needs to be fixed. Basic HTML was added to the English and Profile pages.