V11: 15/6/25
Continued to work on the HTML and CSS  for all the pages to make sure that all the links for the navigation bar were updated to make sure they all work. I also restored the old homepage which might replace the current homepage I can make the navigation bar work properly as the old navigation bar is more aesthetically pleasing which promotes a better user experience.