V17-29/6/25-30/6/25
Added the new navigation bar design to the other pages and changed the relevant details such as the HTML, CSS, and JS for each section. I also changed the colour scheme to better suite the purpose of the website as it is now a shade of maroon which matches the school colours.